pipeline/pipeline/requirements.txt
requests>=2.31.0
huggingface_hub>=0.20.0
ddgs>=9.0.0
googlesearch-python>=1.2.0
python-dotenv>=1.0.0
tldextract>=5.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
textstat>=0.7.3
python-whois>=0.9.0
doubleml>=0.7
lightgbm>=4.0
scikit-learn>=1.3
matplotlib>=3.7
pandas>=2.0
numpy>=1.24
sentence-transformers>=2.2.0
rank_bm25>=0.2.2
scipy>=1.10
